+++ src/library/adash-errors.adb
@@ -48,6 +48,7 @@ package body Adash.Errors is
             | Error_Directory_Not_Found
             | Error_Directory_Denied
             | Error_Not_A_Directory
+            | Error_Handler_Not_A_Name
             | Error_Command_Wrong_Arguments
             | Error_Command_Unavailable
             | Error_No_History_Here
@@ -243,6 +244,8 @@ package body Adash.Errors is
          when Error_Directory_Denied       => return Adash.Messages.Msg_Directory_Denied;
          when Error_Not_A_Directory        =>
             return Adash.Messages.Msg_Not_A_Directory;
+         when Error_Handler_Not_A_Name     =>
+            return Adash.Messages.Msg_Handler_Not_A_Name;
          when Error_Source_Unreadable      => return Adash.Messages.Msg_Source_Unreadable;
          when Error_Source_Too_Large       => return Adash.Messages.Msg_Source_Too_Large;
          when Error_Module_Not_Found       =>